add a delete function to the menu operation.

+++ pointofsale/routes.py
@@ -31,22 +31,22 @@ def add_menus():
 def edit_menus(menus_id):
     menus = Menus.query.get_or_404(menus_id)
     if request.method == "POST":
-        menus=Menus(
-            menus_name=request.form.get("menus_name"),
-            menus_description=request.form.get("menus_description")
-        )
+        menus.menus_name=request.form.get("menus_name"),
+        menus.menus_description=request.form.get("menus_description")
         db.session.add(menus)
         db.session.commit()
         return redirect(url_for("main_menu"))
     return render_template("edit_menus.html", menus=menus)
 
 
-@app.route("/delete_menus/<int:menus_id>")
+@app.route("/delete_menus/<int:menus_id>", methods=["GET", "POST"])
 def delete_menus(menus_id):
     menus = Menus.query.get_or_404(menus_id)
-    db.session.delete(menus)
-    db.session.commit()
-    return redirect(url_for("main_menu"))
+    if request.method == "POST":
+        db.session.delete(menus)
+        db.session.commit()
+        return redirect(url_for("main_menu"))
+    return render_template("delete_menus.html", menus=menus)
 
 
 @app.route("/add_sub_menus", methods=["GET", "POST"])
